Langchain_flask/requirements.txt
flask 
langchain 
openai 
python-dotenv
google-generativeai
